refactor: extract active_transport into its own file
